+++ core/src/main/java/io/questdb/cairo/wal/WalWriter.java
@@ -83,7 +83,7 @@ public class WalWriter implements TableWriterFrontend {
     private int columnCount;
     private long currentTxnStartRowNum = -1;
     private long rowCount = -1;
-    private long segmentId = -1;
+    private int segmentId = -1;
     private long segmentStartMillis;
     private boolean txnOutOfOrder = false;
     private long txnMinTimestamp = Long.MAX_VALUE;
@@ -326,7 +326,7 @@ public class WalWriter implements TableWriterFrontend {
         }
     }
 
-    public long getSegmentId() {
+    public int getSegmentId() {
         return segmentId;
     }
 
@@ -391,7 +391,7 @@ public class WalWriter implements TableWriterFrontend {
 
     public void rollUncommittedToNewSegment() {
         final long uncommittedRows = rowCount - currentTxnStartRowNum;
-        final long newSegmentId = segmentId + 1;
+        final int newSegmentId = segmentId + 1;
 
         path.trimTo(rootLen);
 
@@ -807,7 +807,7 @@ public class WalWriter implements TableWriterFrontend {
         }
     }
 
-    private int createSegmentDir(long segmentId) {
+    private int createSegmentDir(int segmentId) {
         path.trimTo(rootLen);
         path.slash().put(segmentId);
         final int segmentPathLen = path.length();
@@ -1017,7 +1017,7 @@ public class WalWriter implements TableWriterFrontend {
         events.startTxn();
     }
 
-    private void rollLastWalEventRecord(long newSegmentId, long uncommittedRows) {
+    private void rollLastWalEventRecord(int newSegmentId, long uncommittedRows) {
         events.rollback();
         path.trimTo(rootLen).slash().put(newSegmentId);
         events.openEventFile(path, path.length());
